Move environment management outside of project
Since R's build command copies the entire contents of the directory to a
temporary directory then removes items matching the `.Rbuildignore` rules, the
entirety of the direnv cache directory gets copied each time the project is
built or checked. The contents of direnv includes links that are followed and
copied, maintaining their read-only file attributes. This copy both slows down
the process and causes failures due to R's inability to delete the read-only
files.

To solve, direnv related files are moved to a dummy directory above the current
project.

The alternative solution is ideally having R's build system only copy the files
that are allowed by `.Rbuildignore` in the first place rather than recursively
copying everything first and pruning afterwards.