Remove linkingTo dependency on Matrix
The handling of sparse matrices from the `Matrix` package was messy and
was dependent on the newer `Matrix` API which the version of R shipped with
fedora does not yet support.

Removes Matrix specific code by instead passing the matrix needed attributes to
the C code as normal C arrays.

This also removes the need of passing SEXPs to C allowing the use of the
simpler `.C` call instead of requiring `.Call`.

No comparison of performance was done. It seems unlikely to have a significant
impact but may be worth looking at when finding performance bottlenecks in
larger graphs.