Simplify workflows to remove release
- Remove use of `R CMD config C` for finding C compilers to pass to cmake.
Instead let cmake determine the correct C/CXX compilers to use. The R command
can return flags along with the compiler that confuse cmake. It may be
necessary to add this back in later if cmake and R find different compilers to
use in some environments but for now this seems like the simplest approach.