Copy arpack files instead of using submodule
Since the arpack files needed to be modified, in order to use them as a
submodule I would need to make a public clone. I don't want to do this as my
changes are unlikely to be beneficial to anyone else so copying the needed and
modified files directly into this repo seems like the best approach.