Convert directly from R SEXP types -> SE2 neigh list type
- No longer go through the igraph graph type first. Because this type stores
data in a coordinate style sparse matrix (from, to, weight for every edge),
when using large full weighted matrices (such as a correlation matrix) this can
take up a multiple of the original data and be slow to construct. Constructing
the neighbor list directly can be performed much faster and take advantage of
implicit information about the node pairs of every edge.